# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APEX INVESTOR API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardDataSectionCountsItem = {
  slug: string;
  name: string;
  count: number;
};

export type EntryMetadata = { [key: string]: unknown } | null;

export interface Entry {
  id: number;
  sectionId: number;
  sectionSlug?: string;
  sectionName?: string;
  title: string;
  content?: string | null;
  tags: string[];
  starred: boolean;
  archived: boolean;
  metadata?: EntryMetadata;
  createdAt: string;
  updatedAt: string;
}

export interface PortfolioSummary {
  totalValue: number;
  totalCost: number;
  totalUnrealisedPnl: number;
  totalUnrealisedPnlPercent: number;
  holdingCount: number;
  bestPerformer?: string | null;
  worstPerformer?: string | null;
}

export interface DashboardData {
  sectionCounts: DashboardDataSectionCountsItem[];
  recentEntries: Entry[];
  starredEntries: Entry[];
  totalEntries: number;
  portfolioSummary?: PortfolioSummary;
}

export type SearchResultType =
  (typeof SearchResultType)[keyof typeof SearchResultType];

export const SearchResultType = {
  entry: "entry",
  portfolio: "portfolio",
  watchlist: "watchlist",
} as const;

export interface SearchResult {
  id: number;
  sectionSlug: string;
  sectionName: string;
  title: string;
  excerpt?: string;
  type: SearchResultType;
  createdAt: string;
}

export interface Tag {
  id: number;
  name: string;
  color: string;
}

export interface CreateTag {
  name: string;
  color?: string;
}

export interface Section {
  id: number;
  name: string;
  slug: string;
  category: string;
  icon: string;
  description?: string;
  sortOrder: number;
  entryCount?: number;
}

export type CreateEntryMetadata = { [key: string]: unknown } | null;

export interface CreateEntry {
  sectionSlug: string;
  title: string;
  content?: string | null;
  tags?: string[];
  starred?: boolean;
  metadata?: CreateEntryMetadata;
}

export type UpdateEntryMetadata = { [key: string]: unknown } | null;

export interface UpdateEntry {
  title?: string;
  content?: string | null;
  tags?: string[];
  starred?: boolean;
  archived?: boolean;
  metadata?: UpdateEntryMetadata;
}

export type PortfolioItemStatus =
  (typeof PortfolioItemStatus)[keyof typeof PortfolioItemStatus];

export const PortfolioItemStatus = {
  holding: "holding",
  sold: "sold",
  watching: "watching",
} as const;

export interface PortfolioItem {
  id: number;
  ticker: string;
  companyName: string;
  entryPrice: number;
  currentPrice?: number | null;
  shares: number;
  entryDate?: string | null;
  exitDate?: string | null;
  exitPrice?: number | null;
  status: PortfolioItemStatus;
  entryReason?: string | null;
  holdingReason?: string | null;
  sellReason?: string | null;
  sector?: string | null;
  country?: string | null;
  notes?: string | null;
  unrealisedPnl?: number | null;
  unrealisedPnlPercent?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreatePortfolioItemStatus =
  (typeof CreatePortfolioItemStatus)[keyof typeof CreatePortfolioItemStatus];

export const CreatePortfolioItemStatus = {
  holding: "holding",
  sold: "sold",
  watching: "watching",
} as const;

export interface CreatePortfolioItem {
  ticker: string;
  companyName: string;
  entryPrice: number;
  currentPrice?: number | null;
  shares: number;
  entryDate?: string | null;
  status: CreatePortfolioItemStatus;
  entryReason?: string | null;
  holdingReason?: string | null;
  sellReason?: string | null;
  sector?: string | null;
  country?: string | null;
  notes?: string | null;
}

export type UpdatePortfolioItemStatus =
  (typeof UpdatePortfolioItemStatus)[keyof typeof UpdatePortfolioItemStatus];

export const UpdatePortfolioItemStatus = {
  holding: "holding",
  sold: "sold",
  watching: "watching",
} as const;

export interface UpdatePortfolioItem {
  ticker?: string;
  companyName?: string;
  entryPrice?: number;
  currentPrice?: number | null;
  shares?: number;
  entryDate?: string | null;
  exitDate?: string | null;
  exitPrice?: number | null;
  status?: UpdatePortfolioItemStatus;
  entryReason?: string | null;
  holdingReason?: string | null;
  sellReason?: string | null;
  sector?: string | null;
  country?: string | null;
  notes?: string | null;
}

export interface WatchlistItem {
  id: number;
  ticker: string;
  companyName: string;
  sector?: string | null;
  country?: string | null;
  addedDate?: string | null;
  reason?: string | null;
  targetPrice?: number | null;
  currentPrice?: number | null;
  status: string;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateWatchlistItem {
  ticker: string;
  companyName: string;
  sector?: string | null;
  country?: string | null;
  reason?: string | null;
  targetPrice?: number | null;
  currentPrice?: number | null;
  status: string;
  notes?: string | null;
}

export interface UpdateWatchlistItem {
  ticker?: string;
  companyName?: string;
  sector?: string | null;
  country?: string | null;
  reason?: string | null;
  targetPrice?: number | null;
  currentPrice?: number | null;
  status?: string;
  notes?: string | null;
}

export type GlobalSearchParams = {
  q: string;
  sectionSlug?: string;
};

export type GetEntriesParams = {
  sectionSlug: string;
  archived?: boolean;
  starred?: boolean;
  tag?: string;
};

export type GetRecentEntriesParams = {
  limit?: number;
};

export type GetPortfolioParams = {
  status?: GetPortfolioStatus;
};

export type GetPortfolioStatus =
  (typeof GetPortfolioStatus)[keyof typeof GetPortfolioStatus];

export const GetPortfolioStatus = {
  holding: "holding",
  sold: "sold",
  watching: "watching",
} as const;

export type GetWatchlistParams = {
  status?: string;
  sector?: string;
};