# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: APEX INVESTOR API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tags:
    get:
      operationId: getTags
      tags: [tags]
      summary: Get all tags
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTag"
      responses:
        "201":
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /tags/{id}:
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete a tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /entries:
    get:
      operationId: getEntries
      tags: [entries]
      summary: Get entries for a section
      parameters:
        - name: sectionSlug
          in: query
          required: true
          schema:
            type: string
        - name: archived
          in: query
          required: false
          schema:
            type: boolean
        - name: starred
          in: query
          required: false
          schema:
            type: boolean
        - name: tag
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
    post:
      operationId: createEntry
      tags: [entries]
      summary: Create an entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntry"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /entries/recent:
    get:
      operationId: getRecentEntries
      tags: [entries]
      summary: Get recent entries across all sections
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
...
  /entries/{id}:
    get:
      operationId: getEntry
      tags: [entries]
      summary: Get a single entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Not found
    put:
      operationId: updateEntry
      tags: [entries]
      summary: Update an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntry"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /entries/{id}/star:
    patch:
      operationId: toggleStar
      tags: [entries]
      summary: Toggle star on entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /entries/{id}/archive:
    patch:
      operationId: toggleArchive
      tags: [entries]
      summary: Toggle archive on entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /portfolio:
    get:
      operationId: getPortfolio
      tags: [portfolio]
      summary: Get portfolio entries
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [holding, sold, watching]
      responses:
        "200":
          description: Portfolio items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioItem"
    post:
      operationId: createPortfolioItem
      tags: [portfolio]
      summary: Create portfolio item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioItem"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
...
  /portfolio/{id}:
    put:
      operationId: updatePortfolioItem
      tags: [portfolio]
      summary: Update portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePortfolioItem"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
    delete:
      operationId: deletePortfolioItem
      tags: [portfolio]
      summary: Delete portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
      required: [id, name, color]
    CreateTag:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
      required: [name]
...
    Entry:
      type: object
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        sectionSlug:
          type: string
        sectionName:
          type: string
        title:
          type: string
        content:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        starred:
          type: boolean
        archived:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sectionId
        - title
        - tags
        - starred
        - archived
        - createdAt
        - updatedAt
    CreateEntry:
      type: object
      properties:
        sectionSlug:
          type: string
        title:
          type: string
        content:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        starred:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
          nullable: true
      required: [sectionSlug, title]
    UpdateEntry:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        starred:
          type: boolean
        archived:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    PortfolioItem:
      type: object
      properties:
        id:
          type: integer
        ticker:
          type: string
        companyName:
          type: string
        entryPrice:
          type: number
        currentPrice:
          type: number
          nullable: true
        shares:
          type: number
        entryDate:
          type: string
          nullable: true
        exitDate:
          type: string
          nullable: true
        exitPrice:
          type: number
          nullable: true
        status:
          type: string
          enum: [holding, sold, watching]
        entryReason:
          type: string
          nullable: true
        holdingReason:
          type: string
          nullable: true
        sellReason:
          type: string
          nullable: true
        sector:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        unrealisedPnl:
          type: number
          nullable: true
        unrealisedPnlPercent:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ticker
        - companyName
        - entryPrice
        - shares
        - status
        - createdAt
        - updatedAt
    CreatePortfolioItem:
      type: object
      properties:
        ticker:
          type: string
        companyName:
          type: string
        entryPrice:
          type: number
        currentPrice:
          type: number
          nullable: true
        shares:
          type: number
        entryDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [holding, sold, watching]
        entryReason:
          type: string
          nullable: true
        holdingReason:
          type: string
          nullable: true
        sellReason:
          type: string
          nullable: true
        sector:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [ticker, companyName, entryPrice, shares, status]
    UpdatePortfolioItem:
      type: object
      properties:
        ticker:
          type: string
        companyName:
          type: string
        entryPrice:
          type: number
        currentPrice:
          type: number
          nullable: true
        shares:
          type: number
        entryDate:
          type: string
          nullable: true
        exitDate:
          type: string
          nullable: true
        exitPrice:
          type: number
          nullable: true
        status:
          type: string
          enum: [holding, sold, watching]
        entryReason:
          type: string
          nullable: true
        holdingReason:
          type: string
          nullable: true
        sellReason:
          type: string
          nullable: true
        sector:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...