# lib/api-zod/src/generated/types/updateEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APEX INVESTOR API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEntryMetadata } from "./updateEntryMetadata";

export interface UpdateEntry {
  title?: string;
  content?: string | null;
  tags?: string[];
  starred?: boolean;
  archived?: boolean;
  metadata?: UpdateEntryMetadata;
}