# lib/api-zod/src/generated/types/updatePortfolioItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APEX INVESTOR API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePortfolioItemStatus } from "./updatePortfolioItemStatus";

export interface UpdatePortfolioItem {
  ticker?: string;
  companyName?: string;
  entryPrice?: number;
  currentPrice?: number | null;
  shares?: number;
  entryDate?: string | null;
  exitDate?: string | null;
  exitPrice?: number | null;
  status?: UpdatePortfolioItemStatus;
  entryReason?: string | null;
  holdingReason?: string | null;
  sellReason?: string | null;
  sector?: string | null;
  country?: string | null;
  notes?: string | null;
}